Log Errors to Orgfile
Because we run our memacs modules with cronjob we have no chance to get informed in case of an error.
This patch writes Org-Entrys to outputfile-path/error.org_archive -> we can include this file in org-agenda-view to see those errors.
every string from logging.error() is written to error file

updated main:

if an unhandled Exception is thrown this will be also logged with logging.error() (and this will be in the error file :)

Signed-off-by: Armin Wieser <armin.wieser@gmail.com>